Floss: Fix a potential arithmetic overflow
The write_block is set based on the frame_rate which is determined by
the sample_rate we support. In Floss a2dp case, we limit the supported
sample rate so this should be protected by it and shouldn't bring us
actual issue.

Still nice to fix it.

BUG=None
TEST=build

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4541661
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Auto-Submit: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>

+++ cras/src/server/cras_fl_pcm_iodev.c
@@ -311,8 +311,8 @@ static int a2dp_configure_dev(struct cras_iodev* iodev) {
    * And make buffer_size integer multiple of write_block so we
    * don't get cut easily in ring buffer. */
   a2dpio->write_block = iodev->format->frame_rate * PCM_BLOCK_MS / 1000;
-  iodev->buffer_size =
-      PCM_BUF_MAX_SIZE_FRAMES / a2dpio->write_block * a2dpio->write_block;
+  iodev->buffer_size = (unsigned long)PCM_BUF_MAX_SIZE_FRAMES /
+                       a2dpio->write_block * a2dpio->write_block;
 
   a2dpio->pcm_buf = byte_buffer_create(iodev->buffer_size * format_bytes);
   if (!a2dpio->pcm_buf) {
